#309: Fixed bad casting
Task-Url: http://www.trac-wsbc.linkpc.net:8080/trac/QuimP/ticket/309

+++ src/main/java/uk/ac/warwick/wsbc/QuimP/plugin/randomwalk/RandomWalkSegmentationPlugin_.java
@@ -503,8 +503,9 @@ public class RandomWalkSegmentationPlugin_ implements PlugIn, ActionListener, Ch
                     8e-3 // error
             );
             //!<
-            shrinkPower = (Double) sShrinkPower.getValue(); // shrinking object
-            expandPower = (Double) sExpandPower.getValue(); // expanding to get background
+            shrinkPower = ((Integer) sShrinkPower.getValue()).doubleValue(); // shrinking object
+            expandPower = ((Integer) sExpandPower.getValue()).doubleValue(); // expanding to get
+                                                                             // background
             // all ok - store images to later use
             image = tmpImage;
             seedImage = tmpSeed;
